% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstanlm-package.R
\docType{package}
\name{rstanlm-package}
\alias{rstanlm-package}
\alias{rstanlm}
\title{The 'rstanlm' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.3. https://mc-stan.org
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
